Organize test cases
 - Move unit test cases that existed in IntegrationTest class
 - Change the name of IntegrationTest
 - Move serial number pattern test to unit test class

Reduce the number of duplicate integration tests.
