+++ jpsonic-main/src/test/java/org/airsonic/player/service/upnp/processor/ArtistUpnpProcessorTestCase.java
@@ -35,7 +35,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.tesshu.jpsonic.domain.SortingIntegrationTestCase.jPSonicNaturalList;
+import static com.tesshu.jpsonic.domain.JpsonicComparatorsIntegrationTest.jPSonicNaturalList;
 import static org.junit.Assert.assertEquals;
 
 /*
