+++ jpsonic-main/src/test/java/org/airsonic/player/service/upnp/processor/PlaylistUpnpProcessorTestCase.java
@@ -18,7 +18,7 @@
  */
 package org.airsonic.player.service.upnp.processor;
 
-import com.tesshu.jpsonic.domain.SortingIntegrationTestCase;
+import com.tesshu.jpsonic.domain.JpsonicComparatorsIntegrationTest;
 import org.airsonic.player.dao.MediaFileDao;
 import org.airsonic.player.dao.PlaylistDao;
 import org.airsonic.player.domain.Album;
@@ -94,7 +94,7 @@ public class PlaylistUpnpProcessorTestCase extends AbstractAirsonicHomeTest {
 
         if (0 == playlistDao.getAllPlaylists().size()) {
             List<String> shallow = new ArrayList<>();
-            shallow.addAll(SortingIntegrationTestCase.jPSonicNaturalList);
+            shallow.addAll(JpsonicComparatorsIntegrationTest.jPSonicNaturalList);
             Collections.shuffle(shallow);
             shallow.stream().map(toPlaylist).forEach(p -> playlistDao.createPlaylist(p));
         }
